+++ webview-ui/src/components/Announcement.tsx
@@ -32,11 +32,10 @@ const Announcement = ({ version, hideAnnouncement }: AnnouncementProps) => {
 				🎉{"  "}New in v{version}
 			</h3>
 			<p style={{ margin: "5px 0px" }}>
-				Commands now run directly in your terminal thanks to VSCode 1.93's new shell integration updates! Plus a
-				new 'Proceed While Running' button to let Claude continue working while commands run, sending him new
-				output along the way (i.e. letting him react to server errors as he edits files).{" "}
+				Claude can now monitor workspace problems to keep updated on linter/compiler/build issues, letting him
+				proactively fix errors on his own! (adding missing imports, fixing type errors, etc.)
 				<VSCodeLink style={{ display: "inline" }} href="https://x.com/sdrzn/status/1833316974518014072">
-					Demo here!
+					See a demo here.
 				</VSCodeLink>
 			</p>
 			{/*<ul style={{ margin: "0 0 8px", paddingLeft: "12px" }}>